receiver file added

+++ pointCloud.py
@@ -84,7 +84,7 @@ try:
 
                 mask = mask.astype(np.int)
                 asparagusMask = np.where(mask == 1)
-                contours, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+                contours, _ = cv2.findContours(mask.astype(np.uint8), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
 
                 for contour in contours:
                     epsilon = 0.02 * cv2.arcLength(contour, True)
